+++ lib/references/library/sharing.ts
@@ -1,9 +1,10 @@
-'use server'
-
 /**
  * Share-/Portfolio-API (Barrel). Implementierung in sharing-*.ts.
+ *
+ * Kein `'use server'` hier: Turbopack erlaubt in Server-Action-Dateien keine
+ * Re-Exports/`export type` — nur async Function-Exports.
  */
-export type { CreateSharedPortfolioRecipient } from '@/lib/references/library/sharing-create'
+export type { CreateSharedPortfolioRecipient } from '@/lib/references/library/sharing-helpers'
 export {
   createSharedPortfolioImpl,
   getPortfolioManageAndPreviewUrlsForApprovalEmail,
